Added option to select only parent state
This option is only added for parent states that actually
appear in a reaction (i.e. the state is 'valid').
For now the parent state can be selected by selecting the
'-' option in the child select. I am not completely happy
with this UI, as it is quite confusing. A symbol change
might be enough, but a different UI solution is probably
a better option.