Init rebuild of cs bag page in `app` directory
The cross section bag page is currently not working in
production builds. This is due to some code that is
meant to run on the server, spilling onto the client.
The `app` directory allows for a lot more fine grained
control of separation between server and client, thus
it can help to alleviate such problems.
Eventually, all of the `pages` will be updated and
moved to `app`. This commit starts such efforts for
the cross section bag page.