Use production build during tests
Usage of the dev build caused some tests to fail that work
fine in the production environment (which is what we care
about in the end), therefore resulting in false negatives.
This should also greatly speed up the test-app job in the
pipeline.
Revert some previous changes that tried to cover up the
anomalies introduced by the dev build (e.g. needing to
click a button twice upon first page load).