Add `ScientificInput` component
This component uses and internal string buffer to store the
actual input. When the buffer is parsable as a valid (scientific)
number, a float is parsed from the buffer and passe to the parent
via the `onChange`.
It also forbids the entry of characters that are generally invalid
(e.g. arbitrary letters, or a double decimal point).