Add Bolsig+ related components
The `BolsigComponent` component uses the `Bolsig` class to
retrieve and plot the results from a running bolsig server
instance.
The `BolsigInputForm` component builds as form containing
inputs for the properties of the `BolsigInput` type,
required to run a simulation on a running bolsig server
instance.
The `BolsigInput` type is extended to contain a separate
`numerics` section.