Add preliminary `scat-cs/compute` page
This goal of this page is to allow the user to input the
parameters required to run a bolsig simulation on the
selected cross section data. For now it just presents
the different components and can be used to run a
static bolsig simulation. The `BolsigInputForm` and
`BolsigComponent` components still need to be coupled.
Additionally, a ui component needs to be built for the
`composition` section of the Boltzmann input schema.