Update `CSTable` component
Top-level rows will now contain a chevron to indicate that
they can be expanded.
The expanded area will have a bit of padding applied in order
to better distinguish between the expanded area and the next
expandable row.
For the same reason, the custom render of expanded rows has
been reverted (such that they are white, and will thus be
easily distinguishable from the grey expanded area).
Removed the `Id` column from the inner `DataTable`.
The header of the inner tables will no longer highlight on
hover.