Initial pass at automatic form generation from species schemas
Includes form factory functions that use the schema generated from
`AnySpecies` to build an input form for each of the state types.

The method is currently working fairly well, apart from one annoying error
concerning `Select` component that function as a switch between `Single`,
`Compound` and `Unspecified` state descriptors. The error is as follows:
*Warning: A component is changing a controlled input to be uncontrolled.
This is likely caused by the value changing from a defined to undefined,
which should not happen. Decide between using a controlled or uncontrolled
input element for the lifetime of the component.* The result is that the
selected option is not rendered, but the functionality of the component
is fine.

Proper handling of arrays should still be implemented.