Fix typing of the `RouteBuilder`
Splits the former `BaseContext` into two types: `BaseContext`, and `NextContext`.
The former is the return type for the first callback in the chain, i.e. the
callback provided by the init call (thus the init call will return a
`RouteBuilder<BaseContext>` instance). The latter is the first input context to
the callchain, i.e. the context as it is provided by Nextjs.