Render cross section comments on the data inspection page (#2228)
* Move process table classes to dedicated css module

* Render cross section comments

A cross sections comments can be viewed by clicking the corresponding
row. Additionally, a button has been added that shows a tooltip on
hover that indicates that the cross section has comments that can
be viewed.

* Disable row expansion when no comments are present

Also hide the action icon.

* Do not auto open comments when only one process is selected

This would require much more checking of whether the process
has comments. I am also not sure that this is desired behavior
anyway.

* Fix failing end-to-end test