Force the `/api-doc` page to be dynamic
Next would build the `/api-doc` page as a static page,
but the api endpoint URLs rely on `NEXT_PUBLIC_URL`,
which is not generally available at build time. This
is fixed by forcing Next to force the `/api-doc` to
be dynamically rendered (on demand). In terms of
performance this is not a big hit, as the OpenAPI
spec is only generated once and then cached for
reuse.